refactor: 어르신 개인정보, 등록 경로 및 이름 통일
