+++ src/main/java/com/example/medicare_call/controller/action/ElderController.java
@@ -1,17 +1,25 @@
 package com.example.medicare_call.controller.action;
 
 import com.example.medicare_call.domain.Elder;
+import com.example.medicare_call.dto.ElderSettingResponse;
+import com.example.medicare_call.dto.ElderUpdateRequest;
+import com.example.medicare_call.global.annotation.AuthUser;
 import com.example.medicare_call.service.ElderService;
 import com.example.medicare_call.dto.ElderRegisterRequest;
 import com.example.medicare_call.dto.ElderResponse;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @Tag(name = "Elder", description = "어르신(노인) 등록/관리 API")
+@Slf4j
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/elders")
@@ -34,5 +42,47 @@ public class ElderController {
             .guardianName(elder.getGuardian() != null ? elder.getGuardian().getName() : null)
             .build();
         return ResponseEntity.ok(response);
+    }
+
+    @Operation(
+            summary = "어르신 개인정보 조회",
+            description = "로그인한 유저가 등록한 모든 어르신의 개인정보를 조회합니다."
+    )
+    @GetMapping
+    public ResponseEntity<List<ElderSettingResponse>> getElder(@Parameter(hidden = true) @AuthUser Long memberId){
+        log.info("어르신 설정 정보 조회 요청: memberId={}", memberId);
+        List<ElderSettingResponse> body = elderService.getElder(memberId.intValue());
+        return ResponseEntity.ok(body);
+    }
+
+    @Operation(
+            summary = "어르신 개인정보 수정",
+            description = "어르신 정보 전체를 요청받아 값을 수정합니다."
+    )
+    @PostMapping("/{elderId}")
+    public ResponseEntity<ElderSettingResponse> updateElder(
+            @Parameter(hidden = true) @AuthUser Long memberId,
+            @PathVariable Integer elderId,
+            ElderUpdateRequest req
+    ){
+        log.info("어르신 설정 정보 수정 요청: memberId={}, elderId={}, name={}, birthDate={}, gender={}, phone={}, relationship={}, residenceType={}",
+                memberId, elderId, req.name(), req.birthDate(), req.gender(), req.phone(), req.relationship(), req.residenceType());
+
+        ElderSettingResponse res = elderService.updateElder(memberId.intValue(), elderId, req);
+        return ResponseEntity.ok(res);
+    }
+
+    @Operation(
+            summary = "어르신 개인정보 삭제",
+            description = "선택한 어르신의 개인정보를 삭제합니다."
+    )
+    @DeleteMapping("/{elderId}")
+    public ResponseEntity<Void> deleteElder(
+            @Parameter(hidden = true) @AuthUser Long memberId,
+            @PathVariable Integer elderId
+    ){
+        log.info("어르신 설정 정보 삭제 요청: memberId={}, elderId={}", memberId, elderId);
+        elderService.deleteElder(memberId.intValue(), elderId);
+        return ResponseEntity.noContent().build();
     }
 }
